requirements.txt
pyyaml
mypy
numpy
openai
python-dotenv
scipy
matplotlib
seaborn
pandas